PostgreSQL seemed broken
Dropping and recreating the database requires `CREATE DATABASE` permission if you're not a superuser.  Also, the dump defaults to COPY, which cannot be done unless you're a superuser, so it's been changed to use INSERT.

Also, there could be issues with foreign key constraints during the import.  The way around that is to disable triggers, but you have to be a superuser to do that.  Alternatively, you could perform the import as a single transaction, but then any single error will fail the entire operation and you won't know where it happened.

Further, if a database has extensions, those won't propagate during the import, so the rest of the import could fail if anything in the tables uses functionality provided by extensions.  And, again, only a superuser can add extensions to a database.

So, the changes here will require that you add a superuser to your database.yml file, then dumping from one server and importing to another should work as expected.